requirements.txt
requests
beautifulsoup4
PyMuPDF
pdfminer.six
pandas
numpy
yfinance
sentence-transformers
chromadb
openai
python-dotenv
fastapi
uvicorn
aiofiles